+++ lib/hooks/index.ts
@@ -1,4 +1,4 @@
 import useClickedOutside from './useClickedOutside';
-import useFutureGraphData from './useFutureGraphData';
+import useSiteData from './useSiteData';
 
-export { useClickedOutside, useFutureGraphData };
+export { useClickedOutside, useSiteData };
